move service to core lib
