+++ rust/routee-compass-core/src/model/traversal/default/mod.rs
@@ -1,2 +1,3 @@
 pub mod distance_traversal_model;
 pub mod speed_traversal_model;
+pub mod speed_traversal_service;
